requirements.txt
absl-py==0.8.0
appdirs==1.4.3
astor==0.8.0
attrs==19.1.0
beautifulsoup4==4.8.0
black==19.3b0
blis==0.2.4
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.2
fake-useragent==0.1.11
fuzzywuzzy==0.17.0
gast==0.3.2
geographiclib==1.49
geopy==1.20.0
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
idna==2.8
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
matplotlib==3.1.1
murmurhash==1.0.2
mypy==0.720
mypy-extensions==0.4.1
numpy==1.16.0
pandas==0.25.1
parse==1.12.1
plac==0.9.6
preshed==2.0.1
protobuf==3.9.1
PyAudio==0.2.11
pyee==6.0.0
pyparsing==2.4.2
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-html==0.10.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
soupsieve==1.9.3
spacy==2.1.8
SpeechRecognition==3.8.1
srsly==0.1.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
thinc==7.0.8
toml==0.10.0
tqdm==4.35.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.3
w3lib==1.21.0
wasabi==0.2.2
websockets==8.0.2
Werkzeug==0.15.6
wrapt==1.11.2
yahoo-fin==0.8.2
